requirements.txt
langgraph>=0.4.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
tweepy>=4.14.0
